bug(precedence): There was a matching issue with precedence.
There was a matching issue with precedence.  It involved not stopping
when the depth is equivalent to the path length.  This could cause
problems on a very corner case.  Essentially if 2 paths have the same
route except one route is longer with {keys} matcher at the end.

E.G
var router = new Router([{
    route: 'the.force.awakens',
    get: ...
}, {
    route: 'the.force.awakens[{keys}]',
    get: ...
}]);

// This would match both routes since the matching algorithm would not
// stop at the end of the path and `undefined` is a {keys} value.
router.get([['the', 'force', 'awakens']]).subscribe(...);

This issue addresses #162